[feat] create two types of button prefab
1. viewport prefab
2. quickslot prefab
