+++ Assets/Quickslot Button.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a768f9eb5c64b4217b9e8a9b1b0a13a0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
