+++ Assets/Scripts/ManagerTemp.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 426b18a77461d4e88af1a1696524bb6d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
